Rewrite active record integration, vastly simplifying the code and allowing this library to play nicely with other libraries that change the behavior of observers to take arbitrary arguments other than self.
Prior to this change, observable made the assumption that the argument passed to the callback it used to hook into AR
events would always be passed an instance of the object it was observing. This is not the case when some libraries
choose to override this behavior of Rails, such as AttachmentFu. This commit fixes the incompatibility, and also vastly
simplifies the AR integration. This also fixes a bug I didn't track down where my callbacks were being called twice due
to Observable inserting two callbacks for an AR event.